alertsystem/requirements.txt
django
ipdb
Pillow
pytz
